requirements.txt
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.9.0
cryptography==43.0.1
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.7.0
django-filter==24.3
